backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asset-analysis-agent-backend"
version = "0.1.0"
description = "FastAPI + LangChain backend for intelligent asset analysis"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "jinja2>=3.1.4",
  "langchain>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-groq>=0.2.0",
  "pydantic-settings>=2.5.2",
  "psycopg[binary]>=3.2.1",
  "pyyaml>=6.0.2",
  "sqlalchemy>=2.0.35",
  "uvicorn[standard]>=0.30.6",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
app = ["prompts/*.txt", "templates/*.html"]
